refactor(execution/grpc): move execution service where it belongs (#3302)
* refactor(execution/grpc): move execution service where it belongs

* reduce diff

* fix lint
