+++ buf.gen.grpc.yaml
@@ -1,0 +1,11 @@
+version: v2
+
+plugins:
+  - remote: buf.build/protocolbuffers/go
+    out: execution/grpc/types/pb
+    opt: paths=source_relative
+  - remote: buf.build/connectrpc/go
+    out: execution/grpc/types/pb
+    opt: paths=source_relative
+inputs:
+  - directory: execution/grpc/proto
