+++ execution/grpc/handler.go
@@ -10,7 +10,7 @@ import (
 	"golang.org/x/net/http2/h2c"
 
 	"github.com/evstack/ev-node/core/execution"
-	"github.com/evstack/ev-node/types/pb/evnode/v1/v1connect"
+	"github.com/evstack/ev-node/execution/grpc/types/pb/evnode/v1/v1connect"
 )
 
 // NewExecutorServiceHandler creates a new HTTP handler for the ExecutorService.
