+++ execution/grpc/server_test.go
@@ -10,7 +10,7 @@ import (
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/evstack/ev-node/core/execution"
-	pb "github.com/evstack/ev-node/types/pb/evnode/v1"
+	pb "github.com/evstack/ev-node/execution/grpc/types/pb/evnode/v1"
 )
 
 func TestServer_InitChain(t *testing.T) {
